chore(repo): disable less next app
